fix: code split for translators
